mailer/requirements.txt
python-dotenv==1.1.0
pydantic>=2.11.5
pika
